Extract make_method method
